Improves drawable cache memory usage and code
Drawable cache no longer stores constant states that were not part of the
initial preloaded drawables. That way, it's up to the system Resources to
cache the constant states and prune them when the configuration changes.
We also avoid the need to update the dependencies when the configuration
changes, as the "problematic" drawables will have their constant state
recreated.